requirements.txt
anthropic>=0.40.0
openai>=1.60.0
python-dotenv>=1.0.0
markdown>=3.7
